feat: extract super.img

+++ clean_super.sh
@@ -38,6 +38,15 @@ if [[ $? != 0 ]]; then
     exit 1
 fi
 
+echo -e '\e[3;36m[+] Extracting super.img from super.img.lz4\e[0m'
+lz4 -d "$ROOT_DIR/super.img.lz4" "$ROOT_DIR/super.img" &> /dev/null
+if [[ $? != 0 ]]; then
+    echo -e '\e[1;31m[-] Failed to extract the archive\e[0m'
+    exit 1
+fi
+echo -e '\e[3;36m[+] Deleting super.img.lz4 as unneeded\e[0m'
+rm "$ROOT_DIR/super.img.lz4"
+
 echo -e "\e[3;36m[+] Removing working directory\e[0m"
 rm -rf "$ROOT_DIR"
